requirements.txt
keyring==21.4.0
github3.py==1.3.0
